feat: alter table Order

+++ src/main/java/igor/henrique/dtos/order/output/OrderOutputDTO.java
@@ -1,11 +1,13 @@
 package igor.henrique.dtos.order.output;
 
+import java.math.BigDecimal;
+
 public record OrderOutputDTO(
         Long id,
         Integer tableNumber,
         String waiterName,
         String orderStatus,
         String orderDateTime,
-        Double totalPrice
+        BigDecimal totalPrice
 ) {
 }
